+++ qdp-core/src/test/java/com/quantori/qdp/core/task/TaskPersistenceTest.java
@@ -324,8 +324,8 @@ class TaskPersistenceTest extends ContainerizedTest {
     }
 
     @Override
-    public float getPercent() {
-      return 0;
+    public double getPercent() {
+      return 0.0;
     }
   }
 }
